requirements.txt
Flask 
boto3
psycopg2-binary
